feat: initial migration to Apollo 2

+++ AniHyou Widget/AiringWidget.swift
@@ -52,7 +52,7 @@ struct AiringProvider: TimelineProvider {
 
         Network.shared.apollo.fetch(
             query: UserCurrentAnimeListQuery(
-                userId: .some(userId),
+                userId: .some(Int32(userId)),
                 sort: .some([.case(.updatedTimeDesc)])
             ),
             cachePolicy: .fetchIgnoringCacheCompletely
