+++ AniHyou/Views/Activity/ListActivityItemView.swift
@@ -72,7 +72,7 @@ struct ListActivityItemView: View {
                         action: {
                             Task {
                                 if let likeResult = await LikeRepository.toggleLike(
-                                    likeableId: activity.id,
+                                    likeableId: Int32(activity.id),
                                     likeableType: .activity
                                 ) {
                                     isLiked = likeResult
