+++ AniHyou/Views/Activity/TextActivityItemView.swift
@@ -58,7 +58,7 @@ struct TextActivityItemView: View {
                     action: {
                         Task {
                             if let likeResult = await LikeRepository.toggleLike(
-                                likeableId: activity.id,
+                                likeableId: Int32(activity.id),
                                 likeableType: .activity
                             ) {
                                 isLiked = likeResult
